+++ src/integer/mod.rs
@@ -53,7 +53,10 @@ pub use integer::small::SmallInteger;
 
 use std::os::raw::c_int;
 
-#[cfg(try_from)]
+// UGH: two versions of TryFromIntegerError so that the example opens
+// the feature gate if needed (this comment is at the start).
+
+#[cfg(all(try_from, not(nightly_try_from)))]
 /**
 An error which can be returned when a checked conversion from
 [`Integer`] fails.
@@ -79,6 +82,43 @@ println!("Error: {}", error);
 pub struct TryFromIntegerError {
     _unused: (),
 }
+
+// UGH: two versions of TryFromIntegerError so that the example opens
+// the feature gate if needed (this comment is between the versions).
+
+#[cfg(all(try_from, nightly_try_from))]
+/**
+An error which can be returned when a checked conversion from
+[`Integer`] fails.
+
+# Examples
+
+```rust
+# #![feature(try_from)]
+# extern crate rug;
+# fn main() {
+use rug::Integer;
+use rug::integer::TryFromIntegerError;
+use std::convert::TryFrom;
+// This is negative and cannot be converted to u32.
+let i = Integer::from(-5);
+let error: TryFromIntegerError = match u32::try_from(&i) {
+    Ok(_) => unreachable!(),
+    Err(error) => error,
+};
+println!("Error: {}", error);
+# }
+```
+
+[`Integer`]: ../struct.Integer.html
+*/
+#[derive(Clone, Copy, Debug)]
+pub struct TryFromIntegerError {
+    _unused: (),
+}
+
+// UGH: two versions of TryFromIntegerError so that the example opens
+// the feature gate if needed (this comment is at the end).
 
 /**
 The ordering of digits inside a [slice], and bytes inside a digit.