+++ RELEASES.md
@@ -2,8 +2,9 @@ Version 1.2.0 (unreleased)
 ==========================
 
 * Add `Integer::keep_signed_bits` methods.
-* Add `Integer::sum` and `Integer::dot` methods.
-* Add `Float::dot` and `Complex::dot`.
+* Add `Integer::sum` and `Rational::sum` methods.
+* Add `Integer::dot`, `Rational::dot`, `Float::dot` and
+  `Complex::dot`.
 
 Version 1.1.0 (2018-04-23)
 ==========================