+++ RELEASES.md
@@ -1,7 +1,7 @@
 Version 1.2.0 (unreleased)
 ==========================
 
-* Generalize implementations of `Sum` and `Prod` for `Integer` and
+* Generalize implementations of `Sum` and `Product` for `Integer` and
   `Rational`.
 * Add `Integer::keep_signed_bits` methods.
 * Add `Integer::sum` and `Rational::sum` methods.