+++ RELEASES.md
@@ -8,6 +8,11 @@ Version 1.2.0 (unreleased)
 * Add `Integer::dot`, `Rational::dot`, `Float::dot` and
   `Complex::dot`.
 
+Version 1.1.1 (2018-05-20)
+==========================
+
+* Enable unstable `i128`, `u128` and `TryFrom` in nightly.
+
 Version 1.1.0 (2018-04-23)
 ==========================